+++ src/feature/CardRegister/components/CardPreviewSection/CardPreviewSection.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import type { CardPreviewInfoType } from '../../../../common/types/CardPreviewInfoType';
+import type { CardPreviewInfoType } from '../../types/CardPreviewInfoType';
 import CardPreviewContainer from './CardPreviewContainer';
 
 const CardPreviewSection = ({
