+++ src/feature/CardRegisterComplete/CardRegisterCompletePage.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import Button from '../../common/components/Button';
 import checkIconImg from './assets/Group 54.png';
 import { Navigate, useLocation, useNavigate } from 'react-router-dom';
-import type { CardFormInfoType } from '../../common/types/CardPreviewInfoType';
+import type { CardFormInfoType } from '../CardRegister/types/CardPreviewInfoType';
 import { getCardCompanyName } from '../CardRegister/utils/cardDisplay';
 
 const CardRegisterCompletePage = () => {
